refactored to separate backend and frontend
