+++ backend/package.json
@@ -3,13 +3,6 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.10.6",
-    "@emotion/styled": "^11.10.6",
-    "@mui/material": "^5.11.10",
-    "@reduxjs/toolkit": "^1.9.5",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^13.5.0",
     "aws-sdk": "^2.1415.0",
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.1",
@@ -26,20 +19,6 @@
     "multer-s3": "^2.10.0",
     "node-pre-gyp": "^0.17.0",
     "nodemon": "^2.0.20",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-easy-crop": "^4.7.1",
-    "react-icons": "^4.4.0",
-    "react-redux": "^8.1.1",
-    "react-router-dom": "^6.3.0",
-    "react-scripts": "^5.0.1",
-    "react-slider": "^2.0.4",
-    "react-textarea-autosize": "^8.5.2",
-    "redux": "^4.2.0",
-    "redux-devtools-extension": "^2.13.9",
-    "redux-logger": "^3.0.6",
-    "redux-persist": "^6.0.0",
-    "redux-thunk": "^2.4.2",
     "uuidv4": "^6.2.13",
     "web-vitals": "^2.1.4"
   },
@@ -69,9 +48,6 @@
   },
   "devDependencies": {
     "autoprefixer": "^10.4.8",
-    "postcss": "^8.4.14",
-    "prettier": "2.7.1",
-    "react-lorem-ipsum": "^1.4.10",
-    "tailwindcss": "^3.1.7"
+    "prettier": "2.7.1"
   }
 }
