+++ fe_temp/.env
@@ -1,4 +1,4 @@
-REACT_APP_API_URL=http://localhost:3001
-REACT_APP_AWS_ACCESS_KEY=AKIAYEHU5YF46E7EEL5D
-REACT_APP_AWS_SECRET_KEY=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
-REACT_APP_S3_BUCKET_REGION=ap-south-1
+VITE_API_URL=http://localhost:3001
+VITE_AWS_ACCESS_KEY=AKIAYEHU5YF46E7EEL5D
+VITE_AWS_SECRET_KEY=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
+VITE_S3_BUCKET_REGION=ap-south-1
