+++ fe_temp/secret.js
@@ -1,0 +1,5 @@
+module.exports =  _JWT_TOKEN_ = 'd@d,S|T|NCgJ[L)4=l2<$bu&:SHCmV[-083;i_[5&Rws9 ?/I6.>$|f5&5c5a|CI';
+module.exports = MONGODB_URL = 'mongodb+srv://admin:admin123@mastercluster.dxy63ez.mongodb.net/General?retryWrites=true&w=majority';
+module.exports = AWS_ACCESS_KEY = 'AKIAYEHU5YF46E7EEL5D';
+module.exports = AWS_SECRET_ACCESS_KEY= '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';
+module.exports = S3_BUCKET_REGION = 'ap-south-1';
