+++ fe_temp/src/rtk/features/userAuthentication/userAuthenticationSlice.js
@@ -1,7 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 import axios from 'axios';
-const api_url = import.meta.env.REACT_APP_API_URL;
+const api_url = import.meta.env.VITE_API_URL;
 const initialState = {
   isLoggedIn: false,
   _id: '',
